Check Whether an Indirect Path is Valid in a Model

It checks whether an indirect path exists in a model.

Usage

check_path(x, y, m = NULL, fit = NULL, est = NULL)

Arguments

x

Character. The name of predictor at the start of the pathway.

y

Character. The name of the outcome variable at the end of the pathway.

m

A vector of the variable names of the mediators. The pathway goes from the first mediator successively to the last mediator. If NULL, the default, the pathway goes from x to y.

fit

The fit object. Currently only supports a lavaan::lavaan object.

est

The output of lavaan::parameterEstimates(). If NULL, the default, it will be generated from fit. If supplied, fit will ge ignored.

Value

Boolean. TRUE is the path is valid, FALSE if the path is invalid.

Details

It checks whether the path defined by a predictor, an outcome, and sequence of mediators exists in a model.

For example, in the following model in lavaan syntax

m1 ~ x
m2 ~ m1
m3 ~ x
y ~ m2 + m3

This path is valid: `x = "x", y = "y", m = c("m1", "m2")``

This path is invalid: `x = "x", y = "y", m = c("m2")``

This path is also invalid: `x = "x", y = "y", m = c("m1", "m2")``

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

library(lavaan)
#> This is lavaan 0.6-12
#> lavaan is FREE software! Please report any bugs.
data(data_serial_parallel)
dat <- data_serial_parallel
mod <-
"
m11 ~ x + c1 + c2
m12 ~ m11 + x + c1 + c2
m2 ~ x + c1 + c2
y ~ m12 + m2 + m11 + x + c1 + c2
"
fit <- sem(mod, dat,
           meanstructure = TRUE, fixed.x = FALSE)

# The following paths are valid
check_path(x = "x", y = "y", m = c("m11", "m12"), fit = fit)
#> [1] TRUE
check_path(x = "x", y = "y", m = "m2", fit = fit)
#> [1] TRUE
# The following paths are invalid
check_path(x = "x", y = "y", m = c("m11", "m2"), fit = fit)
#> [1] FALSE
check_path(x = "x", y = "y", m = c("m12", "m11"), fit = fit)
#> [1] FALSE