Confidence Interval of Indirect Effect or Conditional Indirect Effect

Return the bootstrap confidence interval of the indirect effect or conditional indirect effect stored in the output of indirect_effect() or cond_indirect().

Usage

# S3 method for indirect
confint(object, parm, level = 0.95, ...)

Arguments

object

The output of indirect_effect() or cond_indirect().

parm

Ignored because the stored object always has only one parameter.

level

The level of confidence, default is .95, returning the 95% confidence interval.

...

Additional arguments. Ignored by the function.

Value

A numeric vector of two elements, the limits of the confidence interval.

Details

It extracts and returns the stored bootstrap confidence interval if available.

See also

Examples

dat <- modmed_x1m3w4y1

# Indirect Effect

library(lavaan)
mod1 <-
"
m1 ~ x
m2 ~ m1
y  ~ m2 + x
"
fit <- sem(mod1, dat,
           meanstructure = TRUE, fixed.x = FALSE,
           se = "none", baseline = FALSE)
# R should be at least 2000 or 5000 in real research.
out1 <- indirect_effect(x = "x", y = "y",
                        m = c("m1", "m2"),
                        fit = fit,
                        boot_ci = TRUE, R = 45, seed = 54151,
                        parallel = FALSE,
                        progress = FALSE)
out1
#> 
#> == Indirect Effect ==
#>                                        
#>  Path:               x -> m1 -> m2 -> y
#>  Indirect Effect     0.064             
#>  95.0% Bootstrap CI: [-0.052 to 0.156] 
#> 
#> Computation Formula:
#>   (b.m1~x)*(b.m2~m1)*(b.y~m2)
#> Computation:
#>   (0.52252)*(0.39883)*(0.30941)
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 45 bootstrap samples.
#> 
#> Coefficients of Component Paths:
#>   Path Coefficient
#>   m1~x       0.523
#>  m2~m1       0.399
#>   y~m2       0.309
#> 
confint(out1)
#>           2.5 %    97.5 %
#> y~x -0.05232493 0.1562506