Confidence Intervals of Indirect Effects or Conditional Indirect Effects

Return the confidence intervals of the conditional indirect effects or conditional effects in the output of cond_indirect_effects().

Usage

# S3 method for cond_indirect_effects
confint(object, parm, level = 0.95, ...)

Arguments

object

The output of cond_indirect_effects().

parm

Ignored. Always returns the bootstrap confidence interval of the effects for all levels stored.

level

The level of confidence, default is .95, returning the 95% confidence interval. Ignored for now and will use the level of the stored intervals.

...

Additional arguments. Ignored by the function.

Details

It extracts and returns the columns for confidence intervals, if available.

See also

Examples

library(lavaan)
dat <- modmed_x1m3w4y1
mod <-
"
m1 ~ x  + w1 + x:w1
m2 ~ m1
y  ~ m2 + x + w4 + m2:w4
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

# Examples for cond_indirect():

# Create levels of w1 and w4
w1levels <- mod_levels("w1", fit = fit)
w1levels
#>                 w1
#> M+1.0SD  1.2280576
#> Mean     0.2589999
#> M-1.0SD -0.7100578
w4levels <- mod_levels("w4", fit = fit)
w4levels
#>                 w4
#> M+1.0SD  1.2087784
#> Mean     0.1532493
#> M-1.0SD -0.9022798
w1w4levels <- merge_mod_levels(w1levels, w4levels)

# Conditional effects from x to m1 when w1 is equal to each of the levels
# R should be at least 2000 or 5000 in real research.
out1 <- suppressWarnings(cond_indirect_effects(x = "x", y = "m1",
                      wlevels = w1levels, fit = fit,
                      boot_ci = TRUE, R = 20, seed = 54151,
                      parallel = FALSE))
confint(out1)
#>             2.5 %    97.5 %
#> M+1.0SD 0.5471000 0.9277342
#> Mean    0.4277304 0.5787751
#> M-1.0SD 0.1590856 0.4261901

# Conditional Indirect effect from x1 through m1 to y,
# when w1 is equal to each of the levels
# R should be at least 2000 or 5000 in real research.
out2 <- suppressWarnings(cond_indirect_effects(x = "x", y = "y", m = c("m1", "m2"),
                      wlevels = w1w4levels, fit = fit,
                      boot_ci = TRUE, R = 20, seed = 54151,
                      parallel = FALSE))
confint(out2)
#>                                2.5 %     97.5 %
#> w1: M+1.0SD; w4: M+1.0SD 0.017865369 0.26780620
#> w1: M+1.0SD; w4: Mean    0.039089693 0.21606349
#> w1: M+1.0SD; w4: M-1.0SD 0.039501231 0.22105942
#> w1: Mean; w4: M+1.0SD    0.012618790 0.18747117
#> w1: Mean; w4: Mean       0.027610100 0.15124995
#> w1: Mean; w4: M-1.0SD    0.033227235 0.15280579
#> w1: M-1.0SD; w4: M+1.0SD 0.007372211 0.12836787
#> w1: M-1.0SD; w4: Mean    0.016130507 0.08643641
#> w1: M-1.0SD; w4: M-1.0SD 0.015140119 0.11279254