Generalized Index of Moderated-Mediation

Compute a generalized version of the index of moderated-mediation proposed by Hayes (2015).

Usage

index_of_mome(output, from = NULL, to = NULL, level = 0.95)

Arguments

output

A cond_indirect_effects-class object: The output of cond_indirect_effects().

from

A row number of output.

to

A row number of output. The change in indirect effects is computed by the change in the level(s) of the moderator(s) from Row from to Row to.

level

The level of confidence for the bootstrap confidence interval. Default is .95.

Value

It returns an index_of_mome-class object. This class has a print method (print.index_of_mome()), a coef method (coef.index_of_mome()), and a confint method (confint.index_of_mome()).

Details

Ths function computes a generalized version of the index of moderated-mediation proposed by Hayes (2015).

The variant proposed by Hayes is the change in the indirect effect along a moderated pathway when a moderator increases by one unit.

The generalized version is the change in the indirect effect along a moderated pathway when a moderator increases by an amount meaningful in a context.

For example, the function can compute the change in the indirect effect when a moderator increases by one standard deviation. We call this variant the standardized index of moderated-mediation.

The function can also compute the change in the standardized indirect effect when a moderator increases by one unit. We call this variant the index of standardized moderated-mediation.

The change in the standardized indirect effect when a moderator increases by one standardized deviation is called the standardized index of standardized moderated-mediation.

This function itself does not set the levels of comparison. The levels to be compared need to be set when calling cond_indirect_effects(). This function extracts required information from the output of cond_indirect_effects().

Functions

index_of_mome: Compute the generalized index of moderated-mediation.

References

Hayes, A. F. (2015). An index and test of linear moderated mediation. Multivariate Behavioral Research, 50(1), 1-22. doi:10.1080/00273171.2014.962683

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

library(lavaan)
dat <- modmed_x1m3w4y1
dat$xw1 <- dat$x * dat$w1
mod <-
"
m1 ~ a * x  + f * w1 + d * xw1
y  ~ b * m1 + cp * x
"
fit <- sem(mod, dat, meanstructure = TRUE, fixed.x = FALSE, se = "none", baseline = FALSE)
est <- parameterEstimates(fit)

# Create levels of w1, the moderators
w1levels <- mod_levels("w1", fit = fit)
w1levels
#>                 w1
#> M+1.0SD  1.2280576
#> Mean     0.2589999
#> M-1.0SD -0.7100578

# Conditional effects from x to y when w1 is equal to each of the levels
boot_out <- fit2boot_out_do_boot(fit, R = 50, seed = 4314)
#> The expected CPU time is 1.7 second(s).
#> Could be faster if 'parallel' set to TRUE.
out <- cond_indirect_effects(x = "x", y = "y", m = "m1",
                      wlevels = w1levels, fit = fit,
                      boot_ci = TRUE, boot_out = boot_out)
out
#> 
#> == Conditional indirect effects ==
#> 
#>  Path: x -> m1 -> y
#>  Conditioned on moderator(s): w1
#>  Moderator(s) represented by: w1
#> 
#>      [w1]   (w1)   ind  CI.lo CI.hi Sig  m1~x  y~m1
#> 1 M+1.0SD  1.228 0.068 -0.660 0.503     0.750 0.091
#> 2 Mean     0.259 0.048 -0.445 0.353     0.523 0.091
#> 3 M-1.0SD -0.710 0.027 -0.231 0.216     0.297 0.091
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 50 samples.
#>  - The 'ind' column shows the indirect effects.
#>  - ‘m1~x’,‘y~m1’ is/are the path coefficient(s) along the path
#>    conditional on the moderators.
#> 
out_ind <- index_of_mome(out, from = 2, to = 3)
out_ind
#> 
#> == Conditional indirect effects ==
#> 
#>  Path: x -> m1 -> y
#>  Conditioned on moderator(s): w1
#>  Moderator(s) represented by: w1
#> 
#>      [w1]   (w1)   ind  CI.lo CI.hi Sig  m1~x  y~m1
#> 1 M-1.0SD -0.710 0.027 -0.231 0.216     0.297 0.091
#> 2 Mean     0.259 0.048 -0.445 0.353     0.523 0.091
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 50 samples.
#>  - The 'ind' column shows the indirect effects.
#>  - ‘m1~x’,‘y~m1’ is/are the path coefficient(s) along the path
#>    conditional on the moderators.
#> 
#> 
#> == Index of Moderated-Mediation ==
#> 
#> Levels: 
#>                     w1
#> To: M-1.0SD -0.7100578
#> From: Mean   0.2589999
#> 
#> Change in Indirect Effect:
#>       x y Change  CI.lo CI.hi
#> Index x y -0.021 -0.153 0.117
#> 
#>  - [CI.lo, CI.hi]: 0.95% percentile confidence interval.
#> 
coef(out_ind)
#>      y~m1~x 
#> -0.02060775 
confint(out_ind)
#>              2.5%     97.5%
#> y~m1~x -0.1530286 0.1169893