Mediation Models By Regression

Simple-to-use functions for fitting regression models and testing indirect effects using just one function.

Usage

q_mediation(
  x,
  y,
  m = NULL,
  cov = NULL,
  data = NULL,
  ci_type = "boot",
  level = 0.95,
  R = 100,
  seed = NULL,
  boot_type = c("perc", "bc"),
  model = NULL
)

q_simple_mediation(
  x,
  y,
  m = NULL,
  cov = NULL,
  data = NULL,
  ci_type = "boot",
  level = 0.95,
  R = 100,
  seed = NULL,
  boot_type = c("perc", "bc")
)

q_serial_mediation(
  x,
  y,
  m = NULL,
  cov = NULL,
  data = NULL,
  ci_type = "boot",
  level = 0.95,
  R = 100,
  seed = NULL,
  boot_type = c("perc", "bc")
)

q_parallel_mediation(
  x,
  y,
  m = NULL,
  cov = NULL,
  data = NULL,
  ci_type = "boot",
  level = 0.95,
  R = 100,
  seed = NULL,
  boot_type = c("perc", "bc")
)

# S3 method for class 'q_mediation'
print(
  x,
  digits = 4,
  annotation = TRUE,
  pvalue = TRUE,
  pvalue_digits = 4,
  se = FALSE,
  for_each_path = FALSE,
  se_ci = TRUE,
  wrap_computation = TRUE,
  ...
)

Arguments

x

For q_mediation(), q_simple_mediation(), q_serial_mediation(), and q_parallel_mediation(), it is the name of the predictor. For the print method of these functions, x is the output of these functions.

y

The name of the outcome.

m

A character vector of the name(s) of the mediator(s). For a simple mediation model, it must has only one name. For serial and parallel mediation models, it can have one or more names. For a serial mediation models, the direction of the paths go from the first names to the last names. For example, c("m1", "m3", "m4") denoted that the path is m1 -> m3 -> m4.

cov

The names of the covariates, if any. If it is a character vector, then the outcome (y) and all mediators (m) regress on all the covariates. If it is a named list of character vectors, then the covariates in an element predict only the variable with the name of this element. For example, list(m1 = "c1", dv = c("c2", "c3")) indicates that c1 predicts "m1", while c2 and c3 predicts "dv". Default is NULL, no covariates.

data

The data frame. Note that listwise deletion will be used and only cases with no missing data on all variables in the model (e.g., x, m, y and cov) will be retained.

ci_type

The type of the confidence interval. Only "boot", the default, is supported and recommended for models fitted by regression. Therefore, this argument can be omitted for now. Set it to NULL and confidence intervals are not formed.

level

The level of confidence of the confidence interval. Default is .95 (for 95% confidence intervals).

R

The number of bootstrap samples. Default is 100. Should be set to 5000 or at least 10000.

seed

The seed for the random number generator. Default is NULL. Should nearly always be set to an integer to make the results reproducible.

boot_type

The type of the bootstrap confidence intervals. Default is "perc", percentile confidence interval. Set "bc" for bias-corrected confidence interval.

model

The type of model. For q_mediation(), it can be "simple" (simple mediation model), "serial" (serial mediation model), or "parallel" (parallel mediation model). It is recommended to call the corresponding wrappers directly (q_simple_mediation(), q_serial_mediation(), and q_parallel_mediation()) instead of call q_mediation().

digits

Number of digits to display. Default is 4.

annotation

Logical. Whether the annotation after the table of effects is to be printed. Default is TRUE.

pvalue

Logical. If TRUE, asymmetric p-values based on bootstrapping will be printed if available. Default is TRUE.

pvalue_digits

Number of decimal places to display for the p-values. Default is 4.

se

Logical. If TRUE and confidence intervals are available, the standard errors of the estimates are also printed. They are simply the standard deviations of the bootstrap estimates.

for_each_path

Logical. If TRUE, each of the paths will be printed individually, using the print-method of the output of indirect_effect(). Default is FALSE.

se_ci

Logical. If TRUE and confidence interval has not been computed, the function will try to compute them from stored standard error if the original standard error is to be used. Ignored if confidence interval has already been computed. Default is TRUE.

wrap_computation

Logical. If TRUE, the default, long computational symbols and values will be wrapped to fit to the screen width.

...

Other arguments. If for_each_path is TRUE, they will be passed to the print method of the output of indirect_effect(). Ignored otherwise.

Value

The function q_mediation() returns a q_mediation class object, with its print method.

The function q_simple_mediation() returns a q_simple_mediation class object, which is a subclass of q_mediation.

The function q_serial_mediation() returns a q_serial_mediation class object, which is a subclass of q_mediation.

The function q_parallel_mediation() returns a q_parallel_mediation class object, which is a subclass of q_mediation.

Details

The family of "q" (quick) functions are for testing mediation effects in common models. These functions do the following in one single call:

Fit the regression models.

Compute and test all the indirect effects.

They are easy-to-use and are suitable for common models which are not too complicated. For now, there are "q" functions for these models:

A simple mediation: One predictor (x), one mediator (m), one outcome (y), and optionally some control variables (covariates) (q_simple_mediation())

A serial mediation model: One predictor (x), one or more mediators (m), one outcome (y), and optionally some control variables (covariates). The mediators positioned sequentially between x and y (q_serial_mediation()):

x -> m1 -> m2 -> ... -> y

A parallel mediation model: One predictor (x), one or more mediators (m), one outcome (y), and optionally some control variables (covariates). The mediators positioned in parallel between x and y (q_parallel_mediation()):

x -> m1 -> y

x -> m2 -> y

...

Users only need to specify the x, m, and y variables, and covariates or control variables, if any (by cov), and the functions will automatically identify all indirect effects and total effects.

Note that they are not intended to be flexible. For models that are different from these common models, it is recommended to fit the models manually, either by structural equation modelling (e.g., lavaan::sem()) or by regression analysis using stats::lm() or lmhelprs::many_lm(). See https://sfcheung.github.io/manymome/articles/med_lm.html for an illustration on how to compute and test indirect effects for an arbitrary mediation model.

Workflow

This is the workflow of the "q" functions:

Do listwise deletion based on all the variables used in the models.

Generate the regression models based on the variables specified.

Fit all the models by OLS regression using stats::lm().

Call all_indirect_paths() to identify all indirect paths.

Call many_indirect_effects() to compute all indirect effects and form their confidence intervals.

Call total_indirect_effect() to compute the total indirect effect.

Return all the results for printing.

The output of the "q" functions have a print method for printing all the major results.

Notes

Flexibility

The "q" functions are designed to be easy to use. They are not designed to be flexible. For maximum flexibility, fit the models manually and call functions such as indirect_effect() separately. See https://sfcheung.github.io/manymome/articles/med_lm.html for illustrations.

Monte Carlo Confidence Intervals

We do not recommend using Monte Carlo confidence intervals for models fitted by regression because the covariances between parameter estimates are assumed to be zero, which may not be the case in some models. Therefore, the "q" functions do not support Monte Carlo confidence intervals. If Monte Carlo intervals are desired, please fit the model by structural equation modeling using lavaan::sem().

Methods (by generic)

print(q_mediation): The print method of the outputs of q_mediation(), q_simple_mediation(), q_serial_mediation(), and q_parallel_mediation().

Functions

q_mediation(): The general "q" function for common mediation models. Not to be used directly.

q_simple_mediation(): A wrapper of q_mediation() for simple mediation models (a model with only one mediator).

q_serial_mediation(): A wrapper of q_mediation() for serial mediation models.

q_parallel_mediation(): A wrapper of q_mediation() for parallel mediation models.

See also

Examples

# ===== Simple mediation

# Set R to 5000 or 10000 in real studies
out <- q_simple_mediation(x = "x",
                          y = "y",
                          m = "m",
                          cov = c("c2", "c1"),
                          data = data_med,
                          R = 100,
                          seed = 1234)
out
#> 
#> =============== Simple Mediation Model ===============
#> 
#> Call:
#> 
#> q_simple_mediation(x = "x", y = "y", m = "m", cov = c("c2", "c1"), 
#>     data = data_med, R = 100, seed = 1234)
#> 
#> ===================================================
#> |                Basic Information                |
#> ===================================================
#> 
#> Predictor(x): x
#> Outcome(y): y
#> Mediator(s)(m): m
#> Model: Simple Mediation Model
#> 
#> The regression models fitted:
#> 
#> m ~ x + c2 + c1
#> y ~ m + x + c2 + c1
#> 
#> The number of cases included: 100 
#> 
#> ===================================================
#> |               Regression Results                |
#> ===================================================
#> 
#> 
#> Model:
#> m ~ x + c2 + c1
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  9.68941    0.91979  10.534   <2e-16 ***
#> x            0.93469    0.08083  11.563   <2e-16 ***
#> c2          -0.16841    0.10305  -1.634   0.1055    
#> c1           0.19778    0.07678   2.576   0.0115 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> R-square = 0.5981. Adjusted R-square = 0.5855. F(3, 96) = 47.6223, p < .001
#> 
#> Model:
#> y ~ m + x + c2 + c1
#>             Estimate Std. Error t value Pr(>|t|)   
#> (Intercept)   4.4152     3.3016   1.337  0.18432   
#> m             0.7847     0.2495   3.145  0.00222 **
#> x             0.5077     0.3057   1.661  0.10004   
#> c2           -0.1544     0.2554  -0.604  0.54695   
#> c1            0.1405     0.1941   0.724  0.47093   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> R-square = 0.3576. Adjusted R-square = 0.3305. F(4, 95) = 13.2204, p < .001
#> 
#> ===================================================
#> |             Indirect Effect Results             |
#> ===================================================
#> 
#> ==  Indirect Effect(s)   ==
#>                ind  CI.lo  CI.hi Sig pvalue
#> x -> m -> y 0.7334 0.3224 1.3636 Sig 0.0000
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - The 'ind' column shows the indirect effects.
#>  
#> ==  Indirect Effect(s) (x-variable(s) Standardized)  ==
#>                std  CI.lo  CI.hi Sig pvalue
#> x -> m -> y 0.7739 0.3287 1.4360 Sig 0.0000
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - std: The partially standardized indirect effects. 
#>  - x-variable(s) standardized.
#>  
#> ==  Indirect Effect(s) (y-variable(s) Standardized)  ==
#>                std  CI.lo  CI.hi Sig pvalue
#> x -> m -> y 0.2914 0.1351 0.5292 Sig 0.0000
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - std: The partially standardized indirect effects. 
#>  - y-variable(s) standardized.
#>  
#> ==  Indirect Effect(s) (Both x-variable(s) and y-variable(s) Standardized)  ==
#>                std  CI.lo  CI.hi Sig pvalue
#> x -> m -> y 0.3074 0.1352 0.5396 Sig 0.0000
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - std: The standardized indirect effects.
#>  
#> ===================================================
#> |          Total Indirect Effect Results          |
#> ===================================================
#> 
#> == Indirect Effect  ==
#>                                        
#>  Path:               x -> m -> y       
#>  Indirect Effect:    0.7334            
#>  95.0% Bootstrap CI: [0.3224 to 1.3636]
#>  Bootstrap p-value:  0.0000            
#> 
#> Computation Formula:
#>   (b.m~x)*(b.y~m)
#> 
#> Computation:
#>   (0.93469)*(0.78469)
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> Coefficients of Component Paths:
#>  Path Coefficient
#>   m~x      0.9347
#>   y~m      0.7847
#> 
#> 
#> == Indirect Effect (‘x’ Standardized) ==
#>                                        
#>  Path:               x -> m -> y       
#>  Indirect Effect:    0.7739            
#>  95.0% Bootstrap CI: [0.3287 to 1.4360]
#>  Bootstrap p-value:  0.0000            
#> 
#> Computation Formula:
#>   (b.m~x)*(b.y~m)*sd_x
#> 
#> Computation:
#>   (0.93469)*(0.78469)*(1.05515)
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> Coefficients of Component Paths:
#>  Path Coefficient
#>   m~x      0.9347
#>   y~m      0.7847
#> 
#> NOTE:
#> - The effects of the component paths are from the model, not
#>   standardized.
#> 
#> 
#> == Indirect Effect (‘y’ Standardized) ==
#>                                        
#>  Path:               x -> m -> y       
#>  Indirect Effect:    0.2914            
#>  95.0% Bootstrap CI: [0.1351 to 0.5292]
#>  Bootstrap p-value:  0.0000            
#> 
#> Computation Formula:
#>   (b.m~x)*(b.y~m)/sd_y
#> 
#> Computation:
#>   (0.93469)*(0.78469)/(2.51729)
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> Coefficients of Component Paths:
#>  Path Coefficient
#>   m~x      0.9347
#>   y~m      0.7847
#> 
#> NOTE:
#> - The effects of the component paths are from the model, not
#>   standardized.
#> 
#> 
#> == Indirect Effect (Both ‘x’ and ‘y’ Standardized) ==
#>                                        
#>  Path:               x -> m -> y       
#>  Indirect Effect:    0.3074            
#>  95.0% Bootstrap CI: [0.1352 to 0.5396]
#>  Bootstrap p-value:  0.0000            
#> 
#> Computation Formula:
#>   (b.m~x)*(b.y~m)*sd_x/sd_y
#> 
#> Computation:
#>   (0.93469)*(0.78469)*(1.05515)/(2.51729)
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> Coefficients of Component Paths:
#>  Path Coefficient
#>   m~x      0.9347
#>   y~m      0.7847
#> 
#> NOTE:
#> - The effects of the component paths are from the model, not
#>   standardized.
#> 
#> ===================================================
#> |                      Notes                      |
#> ===================================================
#> 
#> - The asymmetric bootstrap value for an effect is the same whether x
#>   and/or y is/are standardized.

# Different control variables for m and y
# out <- q_simple_mediation(x = "x",
#                           y = "y",
#                           m = "m",
#                           cov = list(m = "c1",
#                                      y = c("c1", "c2")),
#                           data = data_med,
#                           R = 100,
#                           seed = 1234)
# out

# ===== Serial mediation

# Set R to 5000 or 10000 in real studies
out <- q_serial_mediation(x = "x",
                          y = "y",
                          m = c("m1", "m2"),
                          cov = c("c2", "c1"),
                          data = data_serial,
                          R = 100,
                          seed = 1234)
out
#> 
#> =============== Serial Mediation Model ===============
#> 
#> Call:
#> 
#> q_serial_mediation(x = "x", y = "y", m = c("m1", "m2"), cov = c("c2", 
#>     "c1"), data = data_serial, R = 100, seed = 1234)
#> 
#> ===================================================
#> |                Basic Information                |
#> ===================================================
#> 
#> Predictor(x): x
#> Outcome(y): y
#> Mediator(s)(m): m1, m2
#> Model: Serial Mediation Model
#> 
#> The regression models fitted:
#> 
#> m1 ~ x + c2 + c1
#> m2 ~ x + m1 + c2 + c1
#> y ~ m1 + m2 + x + c2 + c1
#> 
#> The number of cases included: 100 
#> 
#> ===================================================
#> |               Regression Results                |
#> ===================================================
#> 
#> 
#> Model:
#> m1 ~ x + c2 + c1
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 10.81566    1.14379   9.456 2.20e-15 ***
#> x            0.82244    0.09424   8.727 8.03e-14 ***
#> c2          -0.18886    0.09278  -2.036   0.0446 *  
#> c1           0.17148    0.09070   1.891   0.0617 .  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> R-square = 0.4774. Adjusted R-square = 0.4611. F(3, 96) = 29.2349, p < .001
#> 
#> Model:
#> m2 ~ x + m1 + c2 + c1
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  3.51937    1.58681   2.218  0.02895 *  
#> x           -0.11610    0.12598  -0.922  0.35909    
#> m1           0.42078    0.10188   4.130  7.8e-05 ***
#> c2          -0.16195    0.09460  -1.712  0.09017 .  
#> c1           0.27753    0.09221   3.010  0.00335 ** 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> R-square = 0.3410. Adjusted R-square = 0.3133. F(4, 95) = 12.2904, p < .001
#> 
#> Model:
#> y ~ m1 + m2 + x + c2 + c1
#>             Estimate Std. Error t value Pr(>|t|)  
#> (Intercept)  9.46788    3.60900   2.623   0.0102 *
#> m1          -0.43534    0.24539  -1.774   0.0793 .
#> m2           0.52077    0.22753   2.289   0.0243 *
#> x            0.49285    0.28063   1.756   0.0823 .
#> c2          -0.09604    0.21300  -0.451   0.6531  
#> c1           0.09884    0.21403   0.462   0.6453  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> R-square = 0.0915. Adjusted R-square = 0.0432. F(5, 94) = 1.8934, p = 0.1028
#> 
#> ===================================================
#> |             Indirect Effect Results             |
#> ===================================================
#> 
#> ==  Indirect Effect(s)   ==
#>                        ind   CI.lo  CI.hi Sig pvalue
#> x -> m1 -> m2 -> y  0.1802  0.0376 0.3518 Sig 0.0000
#> x -> m1 -> y       -0.3580 -0.7488 0.0285     0.0600
#> x -> m2 -> y       -0.0605 -0.1963 0.1371     0.4800
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - The 'ind' column shows the indirect effects.
#>  
#> ==  Indirect Effect(s) (x-variable(s) Standardized)  ==
#>                        std   CI.lo  CI.hi Sig pvalue
#> x -> m1 -> m2 -> y  0.1721  0.0331 0.3249 Sig 0.0000
#> x -> m1 -> y       -0.3419 -0.7121 0.0258     0.0600
#> x -> m2 -> y       -0.0577 -0.1864 0.1164     0.4800
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - std: The partially standardized indirect effects. 
#>  - x-variable(s) standardized.
#>  
#> ==  Indirect Effect(s) (y-variable(s) Standardized)  ==
#>                        std   CI.lo  CI.hi Sig pvalue
#> x -> m1 -> m2 -> y  0.0897  0.0197 0.1758 Sig 0.0000
#> x -> m1 -> y       -0.1782 -0.3599 0.0138     0.0600
#> x -> m2 -> y       -0.0301 -0.1045 0.0644     0.4800
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - std: The partially standardized indirect effects. 
#>  - y-variable(s) standardized.
#>  
#> ==  Indirect Effect(s) (Both x-variable(s) and y-variable(s) Standardized)  ==
#>                        std   CI.lo  CI.hi Sig pvalue
#> x -> m1 -> m2 -> y  0.0856  0.0166 0.1610 Sig 0.0000
#> x -> m1 -> y       -0.1701 -0.3414 0.0125     0.0600
#> x -> m2 -> y       -0.0287 -0.0974 0.0562     0.4800
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - std: The standardized indirect effects.
#>  
#> ===================================================
#> |          Total Indirect Effect Results          |
#> ===================================================
#> 
#> == Indirect Effect  ==
#>                                          
#>  Path:                x -> m1 -> m2 -> y 
#>  Path:                x -> m1 -> y       
#>  Path:                x -> m2 -> y       
#>  Function of Effects: -0.2383            
#>  95.0% Bootstrap CI:  [-0.5488 to 0.1154]
#>  Bootstrap p-value:   0.1400             
#> 
#> Computation of the Function of Effects:
#>  ((x->m1->m2->y)
#> +(x->m1->y))
#> +(x->m2->y) 
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> 
#> == Indirect Effect (‘x’ Standardized) ==
#>                                          
#>  Path:                x -> m1 -> m2 -> y 
#>  Path:                x -> m1 -> y       
#>  Path:                x -> m2 -> y       
#>  Function of Effects: -0.2275            
#>  95.0% Bootstrap CI:  [-0.5008 to 0.1069]
#>  Bootstrap p-value:   0.1400             
#> 
#> Computation of the Function of Effects:
#>  ((x->m1->m2->y)
#> +(x->m1->y))
#> +(x->m2->y) 
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> 
#> == Indirect Effect (‘y’ Standardized) ==
#>                                          
#>  Path:                x -> m1 -> m2 -> y 
#>  Path:                x -> m1 -> y       
#>  Path:                x -> m2 -> y       
#>  Function of Effects: -0.1186            
#>  95.0% Bootstrap CI:  [-0.2845 to 0.0568]
#>  Bootstrap p-value:   0.1400             
#> 
#> Computation of the Function of Effects:
#>  ((x->m1->m2->y)
#> +(x->m1->y))
#> +(x->m2->y) 
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> 
#> == Indirect Effect (Both ‘x’ and ‘y’ Standardized) ==
#>                                          
#>  Path:                x -> m1 -> m2 -> y 
#>  Path:                x -> m1 -> y       
#>  Path:                x -> m2 -> y       
#>  Function of Effects: -0.1132            
#>  95.0% Bootstrap CI:  [-0.2680 to 0.0527]
#>  Bootstrap p-value:   0.1400             
#> 
#> Computation of the Function of Effects:
#>  ((x->m1->m2->y)
#> +(x->m1->y))
#> +(x->m2->y) 
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> ===================================================
#> |                      Notes                      |
#> ===================================================
#> 
#> - The asymmetric bootstrap value for an effect is the same whether x
#>   and/or y is/are standardized.

# Different control variables for m and y
# out <- q_serial_mediation(x = "x",
#                           y = "y",
#                           m = c("m1", "m2"),
#                           cov = list(m1 = "c1",
#                                      m2 = c("c2", "c1"),
#                                      y = "c2"),
#                           data = data_serial,
#                           R = 100,
#                           seed = 1234)
# out

# ===== Parallel mediation

out <- q_parallel_mediation(x = "x",
                            y = "y",
                            m = c("m1", "m2"),
                            cov = c("c2", "c1"),
                            data = data_parallel,
                            R = 100,
                            seed = 1234)
out
#> 
#> =============== Parallel Mediation Model ===============
#> 
#> Call:
#> 
#> q_parallel_mediation(x = "x", y = "y", m = c("m1", "m2"), cov = c("c2", 
#>     "c1"), data = data_parallel, R = 100, seed = 1234)
#> 
#> ===================================================
#> |                Basic Information                |
#> ===================================================
#> 
#> Predictor(x): x
#> Outcome(y): y
#> Mediator(s)(m): m1, m2
#> Model: Parallel Mediation Model
#> 
#> The regression models fitted:
#> 
#> m1 ~ x + c2 + c1
#> m2 ~ x + c2 + c1
#> y ~ m1 + m2 + x + c2 + c1
#> 
#> The number of cases included: 100 
#> 
#> ===================================================
#> |               Regression Results                |
#> ===================================================
#> 
#> 
#> Model:
#> m1 ~ x + c2 + c1
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  9.39260    1.40019   6.708 1.37e-09 ***
#> x            0.87694    0.11441   7.665 1.45e-11 ***
#> c2          -0.09293    0.12255  -0.758   0.4501    
#> c1           0.25189    0.11131   2.263   0.0259 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> R-square = 0.4394. Adjusted R-square = 0.4219. F(3, 96) = 25.0816, p < .001
#> 
#> Model:
#> m2 ~ x + c2 + c1
#>             Estimate Std. Error t value Pr(>|t|)   
#> (Intercept)   4.3796     1.3463   3.253  0.00158 **
#> x             0.2968     0.1100   2.698  0.00825 **
#> c2           -0.3125     0.1178  -2.652  0.00937 **
#> c1            0.2746     0.1070   2.566  0.01185 * 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> R-square = 0.2171. Adjusted R-square = 0.1927. F(3, 96) = 8.8753, p < .001
#> 
#> Model:
#> y ~ m1 + m2 + x + c2 + c1
#>             Estimate Std. Error t value Pr(>|t|)  
#> (Intercept)  2.65618    3.17697   0.836   0.4052  
#> m1           0.48642    0.19106   2.546   0.0125 *
#> m2           0.47128    0.19871   2.372   0.0197 *
#> x            0.29112    0.26939   1.081   0.2826  
#> c2          -0.03212    0.23411  -0.137   0.8912  
#> c1          -0.05256    0.21557  -0.244   0.8079  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> R-square = 0.2830. Adjusted R-square = 0.2449. F(5, 94) = 7.4201, p < .001
#> 
#> ===================================================
#> |             Indirect Effect Results             |
#> ===================================================
#> 
#> ==  Indirect Effect(s)   ==
#>                 ind  CI.lo  CI.hi Sig pvalue
#> x -> m1 -> y 0.4266 0.1859 0.7997 Sig 0.0000
#> x -> m2 -> y 0.1399 0.0109 0.3563 Sig 0.0000
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - The 'ind' column shows the indirect effects.
#>  
#> ==  Indirect Effect(s) (x-variable(s) Standardized)  ==
#>                 std  CI.lo  CI.hi Sig pvalue
#> x -> m1 -> y 0.4279 0.1818 0.8064 Sig 0.0000
#> x -> m2 -> y 0.1403 0.0109 0.3644 Sig 0.0000
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - std: The partially standardized indirect effects. 
#>  - x-variable(s) standardized.
#>  
#> ==  Indirect Effect(s) (y-variable(s) Standardized)  ==
#>                 std  CI.lo  CI.hi Sig pvalue
#> x -> m1 -> y 0.1819 0.0786 0.3289 Sig 0.0000
#> x -> m2 -> y 0.0596 0.0048 0.1505 Sig 0.0000
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - std: The partially standardized indirect effects. 
#>  - y-variable(s) standardized.
#>  
#> ==  Indirect Effect(s) (Both x-variable(s) and y-variable(s) Standardized)  ==
#>                 std  CI.lo  CI.hi Sig pvalue
#> x -> m1 -> y 0.1825 0.0772 0.3245 Sig 0.0000
#> x -> m2 -> y 0.0598 0.0048 0.1510 Sig 0.0000
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - [pvalue] are asymmetric bootstrap p-values.
#>  - std: The standardized indirect effects.
#>  
#> ===================================================
#> |          Total Indirect Effect Results          |
#> ===================================================
#> 
#> == Indirect Effect  ==
#>                                         
#>  Path:                x -> m1 -> y      
#>  Path:                x -> m2 -> y      
#>  Function of Effects: 0.5664            
#>  95.0% Bootstrap CI:  [0.2585 to 1.0374]
#>  Bootstrap p-value:   0.0000            
#> 
#> Computation of the Function of Effects:
#>  (x->m1->y)
#> +(x->m2->y) 
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> 
#> == Indirect Effect (‘x’ Standardized) ==
#>                                         
#>  Path:                x -> m1 -> y      
#>  Path:                x -> m2 -> y      
#>  Function of Effects: 0.5682            
#>  95.0% Bootstrap CI:  [0.2569 to 1.0270]
#>  Bootstrap p-value:   0.0000            
#> 
#> Computation of the Function of Effects:
#>  (x->m1->y)
#> +(x->m2->y) 
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> 
#> == Indirect Effect (‘y’ Standardized) ==
#>                                         
#>  Path:                x -> m1 -> y      
#>  Path:                x -> m2 -> y      
#>  Function of Effects: 0.2415            
#>  95.0% Bootstrap CI:  [0.1084 to 0.3976]
#>  Bootstrap p-value:   0.0000            
#> 
#> Computation of the Function of Effects:
#>  (x->m1->y)
#> +(x->m2->y) 
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> 
#> == Indirect Effect (Both ‘x’ and ‘y’ Standardized) ==
#>                                         
#>  Path:                x -> m1 -> y      
#>  Path:                x -> m2 -> y      
#>  Function of Effects: 0.2423            
#>  95.0% Bootstrap CI:  [0.1082 to 0.3971]
#>  Bootstrap p-value:   0.0000            
#> 
#> Computation of the Function of Effects:
#>  (x->m1->y)
#> +(x->m2->y) 
#> 
#> 
#> Percentile confidence interval formed by nonparametric bootstrapping
#> with 100 bootstrap samples.
#> 
#> ===================================================
#> |                      Notes                      |
#> ===================================================
#> 
#> - The asymmetric bootstrap value for an effect is the same whether x
#>   and/or y is/are standardized.
# Different control variables for m and y
# out <- q_parallel_mediation(x = "x",
#                             y = "y",
#                             m = c("m1", "m2"),
#                             cov = list(m1 = "c1",
#                                        m2 = c("c2", "c1"),
#                                        y = "c2"),
#                             data = data_parallel,
#                             R = 100,
#                             seed = 1234)
# out